fixing bfactor

+++ lab6/src/AVLTree.c
@@ -26,7 +26,7 @@ Node* insert(Node* elem, Node* tree) {
 	tree->link[idx] = insert(elem, tree->link[idx]);
 	tree->height = max(height(tree->link[0]), height(tree->link[1])) + 1;
 
-	if ((1 < height(tree)) && (abs(height(tree->link[idx]) - height(tree->link[!idx])) > 1)) {
+	if ((1 < height(tree)) && (abs((int)height(tree->link[idx]) - (int)height(tree->link[!idx])) > 1)) {
 		if (height((tree->link[idx])->link[idx]) > height((tree->link[idx])->link[!idx]))
 			tree = rotTree(tree, idx);
 		else {
